refactor: CreateWorkspaceCard 폼 로직을 model 훅으로 분리
이름 상태, 검증, 제출 처리를 useCreateWorkspace 훅으로 옮기고
컴포넌트는 렌더링만 담당하도록 정리하였습니다.

#204
